Understand optimization for passing value
C++ uses reference and copy elison to optimize passing value. When it
comes to the lifetime extension of reference, things become much crazier
than I think. A reference never refers to a prvalue/xvalue/lvalue, since
they are expressions! It is the result of the expression that is
referred.

So why do I believe that prvalue/xvalue/lvalue is an object for 5 years?
Maybe that's because they are named as VALUE. Well, now this becomes the
worst naming in C++ for me:)